Change image placeholder to encourage a different Image uploader
With Discord CDN links expiring after 24 Hours and Imgurs recent changes, something new will need to be used for File Uploading to make sure the links do not break. FiveManage is a great uploader we use with some resources and could be a good alternate, unless Community CAD Host their own, or we allow file uploads instead of links!